Completed shop and inventory functionality, items get added and validated, currency gets properly deducted, items get placed in inventory, added more purchasable items for different tiers
Data types classes moved to an include-able file.
Finished shop functionality with complete checks for item presence in inventory, proper currency deduction, visual representation of money left after purchase.
Use of sweet alerts for rich UX.

Added more options for profile icons at different tiers.

Basically nearly finished, only two features remaining:
Gacha - allow students to spend 5 copies of an item for a chance to get a random higher tier one.
Inventory - allow a user to select an item from those owned as their main account profile picture.
