+++ contents/includes/shopPage.php
@@ -1,158 +1,6 @@
 <?php
-class item
-{
-    private $UUID;
-    private $sprite;
-    private $name;
-    private $description;
-    private $rarity;
-    private $price;
-    private $effectType;
-    private $effectValue;
-    public function __construct()
-    {
-    }
-        public function setItemUUID($UUID)
-    {
-        $this->UUID = $UUID;
-    }
-    public function setSprite($sprite)
-    {
-        $this->sprite = $sprite;
-    }
-    public function setName($name)
-    {
-        $this->name = $name;
-    }
-    public function setDescription($desc)
-    {
-        $this->description =$desc;       
-    }
-    public function setRarity($rarity)
-    {
-        $this->rarity = $rarity;    
-    }
-    public function setPrice($price)
-    {
-        $this->price = $price;      
-    }
-    public function setEffectType($eT)
-    {
-        $this->effectType = $eT;
-        
-    }
-    public function setEffectValue($eV)
-    {
-        $this->effectValue = $eV;        
-    }
 
-    public function getUUID()
-    {
-        return $this->UUID;
-    }
-    public function getSprite()
-    {
-        return $this->sprite;
-    }
-    public function getName()
-    {
-        return $this->name;
-    }
-    public function getDescription()
-    {
-        return $this->description;       
-    }
-    public function getRarity()
-    {
-        return $this->rarity;    
-    }
-    public function getPrice()
-    {
-        return $this->price;      
-    }
-    public function getEffectType()
-    {
-        return $this->effectType;
-        
-    }
-    public function getEffectValue()
-    {
-        return $this->effectValue;        
-    }
-}
-class shop
-{
-    private $items; //unsorted total list
-    private $commonItems;
-    private $epicItems;
-    private $legendaryItems;
-    public function __construct()
-    {
-        $this->items = Array();
-        $this->commonItems = Array();
-        $this->epicItems = Array();
-        $this->legendaryItems = Array();
-    }
-
-    public function addItem($item)
-    {
-        array_push($this->items,$item);
-    }
-    public function addCommonItem($item)
-    {
-        array_push($this->commonItems,$item);
-    }
-    public function addEpicItem($item)
-    {
-        array_push($this->epicItems,$item);
-    }
-    public function addLegendaryItem($item)
-    {
-        array_push($this->legendaryItems,$item);
-    }
-
-    public function sortRarities()
-    {
-        $this->commonItems =[];
-        $this->epicItems =[];
-        $this->legendaryItems =[]; 
-        
-        for($i=0;$i<sizeof($this->items);$i++)
-        {
-            if($this->items[$i]->getRarity() == "Common")
-            {
-                $this->addCommonItem($this->items[$i]);
-            }else if($this->items[$i]->getRarity() == "Epic")
-            {
-                $this->addEpicItem($this->items[$i]);
-
-            }else if($this->items[$i]->getRarity() == "Legendary")
-            {
-                $this->addLegendaryItem($this->items[$i]);
-
-            }
-        }
-    }
-
-    public function getItems()
-    {
-        return $this->items;
-    }
-    public function getCommonItems()
-    {
-        return $this->commonItems;
-    }
-    public function getEpicItems()
-    {
-        return $this->epicItems;
-    }
-    public function getLegendaryItems()
-    {
-        return $this->legendaryItems;
-    }
-
-}
-
+require_once("./php/MT_classes.php");
 $db= new inquizitiveDB ();
 $shop = new shop();
 $userUUID= htmlspecialchars($_SESSION['userUUID']); 
@@ -356,6 +204,55 @@ if ($result = $db->Query("CALL GetInquizitiveCurrencyByUserUUID(?);", [$userUUID
                     },
                     error: function(xhr, status, error) {
                         //console.log('Error:', error);
+                    }
+                });
+            })
+        })
+
+        $(".shop-item-price").ready(function(){
+            $(".shop-item-price").click(function (){
+                $.ajax({
+                    url: './shop-buy-item', 
+                    type: 'POST',
+                    data: {
+                        item: $(this).attr("item-id"),
+                        quantity: 1
+                    }, 
+                    success: function(response) {
+                        //console.log('Success:', response);
+                        if(response.status == "success")
+                        {
+                            
+                            Swal.fire({
+                            position: "top-end",
+                            icon: "success",
+                            title: response.message,
+                            showConfirmButton: false,
+                            timer: 1500
+                            });
+                            $("#shop-page-nav-currency").text(response.money);
+                        }else if(response.status == "money404")
+                        {
+                            Swal.fire({
+                            position: "top-end",
+                            icon: "error",
+                            title: response.message,
+                            showConfirmButton: false,
+                            timer: 1500
+                            });
+                        }else
+                        {
+                            Swal.fire({
+                            position: "top-end",
+                            icon: "error",
+                            title: response.status,
+                            showConfirmButton: false,
+                            timer: 1500
+                            });
+                        }
+                    },
+                    error: function(xhr, status, error) {
+    
                     }
                 });
             })
