+++ css/shop.css
@@ -149,6 +149,7 @@
     object-fit: cover;
     height: 100%;
     width: 100%;
+    border-radius: 5px;
 
 }
